Moved CLI parsing tests into test_parsing
