+++ tests/test_cli.py
@@ -104,23 +104,3 @@ fandango:ERROR: Only found 0 perfect solutions, instead of the required 10
         self.assertEqual("", err)
         self.assertEqual("", out)
         self.assertEqual(0, code)
-
-    def test_infinity_abc(self):
-        command = shlex.split("fandango parse -f docs/infinity.fan --validate tests/resources/abc.txt")
-        out, err, code = self.run_command(command)
-        self.assertEqual("", err)
-        self.assertEqual("", out)
-        self.assertEqual(0, code)
-
-    def test_infinity_abcabc(self):
-        command = shlex.split("fandango parse -f docs/infinity.fan --validate tests/resources/abcabc.txt")
-        out, err, code = self.run_command(command)
-        self.assertEqual("", err)
-        self.assertEqual("", out)
-        self.assertEqual(0, code)
-
-    def test_infinity_abcd(self):
-        # This should be rejected by the grammar
-        command = shlex.split("fandango parse -f docs/infinity.fan tests/resources/abcd.txt")
-        out, err, code = self.run_command(command)
-        self.assertEqual(1, code)
